One change to logger.cfg to fix an apparent bug in the qualname parameter of the gplLong logger.
A hopefully temporary change to the stageFiles.py so that before a
file is written to xroot, it is first deleted.  This gets around a
problem with xroot that caused an error if an existing file happened
to reside on a "full" xroot server.  Andy and Fabrizio (xroot
developers) have agreed that a future release will change the behavior
file overwriting to avoid this situation.